Make WarningType and WarningMode API
The reason for making those API is to allow implementors of
WarningTypeCommandDescriptionProviders to use them to construct warning
types in a type-safe way.
