+++ org.openhab.binding.zigbee/src/main/java/org/openhab/binding/zigbee/converter/warningdevice/SquawkType.java
@@ -6,11 +6,11 @@
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
  */
-package org.openhab.binding.zigbee.internal.converter.warningdevice;
+package org.openhab.binding.zigbee.converter.warningdevice;
 
 import static java.util.Arrays.stream;
 import static java.util.stream.Collectors.toMap;
-import static org.openhab.binding.zigbee.internal.converter.warningdevice.SoundLevel.HIGH;
+import static org.openhab.binding.zigbee.converter.warningdevice.SoundLevel.HIGH;
 
 import java.util.Map;
