+++ org.openhab.binding.zigbee/src/test/java/org/openhab/binding/zigbee/internal/converter/warningdevice/SquawkTypeTest.java
@@ -11,6 +11,9 @@ package org.openhab.binding.zigbee.internal.converter.warningdevice;
 import static org.junit.Assert.*;
 
 import org.junit.Test;
+import org.openhab.binding.zigbee.converter.warningdevice.SoundLevel;
+import org.openhab.binding.zigbee.converter.warningdevice.SquawkMode;
+import org.openhab.binding.zigbee.converter.warningdevice.SquawkType;
 
 /**
  * Unit tests for {@link SquawkType}.
